Rename storage to firestore
